editing booking rate to Integer

+++ database/migrations/2026_01_15_120402_add_booking_rate_to_bookings_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('bookings', function (Blueprint $table) {
-            $table->decimal('booking_rate')->default('0');
+            $table->integer('booking_rate')->default('0');
         });
     }
